Moved "VS" components in "VS" namespace
