+++ src/Microsoft.VisualStudio.ProjectSystem.CSharp.VS.UnitTests/ProjectSystem/VS/LanguageServices/CSharpCodeDomProviderTests.cs
@@ -2,7 +2,7 @@
 
 using Xunit;
 
-namespace Microsoft.VisualStudio.ProjectSystem.LanguageServices
+namespace Microsoft.VisualStudio.ProjectSystem.VS.LanguageServices
 {
     [ProjectSystemTrait]
     public class CSharpCodeDomProviderTests
