+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/Input/Commands/AbstractOpenProjectDesignerCommand.cs
@@ -4,7 +4,7 @@ using System;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.ProjectSystem.Utilities.Designers;
 
-namespace Microsoft.VisualStudio.ProjectSystem.Input.Commands
+namespace Microsoft.VisualStudio.ProjectSystem.VS.Input.Commands
 {
     internal abstract class AbstractOpenProjectDesignerCommand : AbstractSingleNodeProjectCommand
     {
