+++ src/Microsoft.VisualStudio.ProjectSystem.VisualBasic.VS/ProjectSystem/VS/VisualBasicProjectDesignerPageProvider.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.VisualStudio.ProjectSystem.Utilities;
 using Microsoft.VisualStudio.ProjectSystem.VS;
 
-namespace Microsoft.VisualStudio.ProjectSystem.Designers
+namespace Microsoft.VisualStudio.ProjectSystem.Designers.VS
 {
     /// <summary>
     ///     Provides project designer property pages.
